==> code/4_save_weather_data.R <==
library(chillR)


# This script is to save the weather data and to delete the .RData files

# This will save as csv files all the dataframes generated by the script "1_weather_data.R"

save_temperature_scenarios(c(SA_GSOD_list, CL_list_90, AR_list_90),
                           path = "./data/original_temps", prefix = "WS")


# Compress the files in order to reduce the size of the repository (a = add to archive, idq = enable quiet mode,
# ep1 = exclude base directory). Note: "zip" parameter should be the directory of the programm to compress the data 

zip(zipfile = "./data/original_temps", files = "./data/original_temps", flags = " a -idq -ep1",
    zip = "C:/Program Files/WinRAR/WinRAR")


# Delete the csv files from the repository to save space

unlink("./data/original_temps", recursive = T)



# Save the weather data generated by patching and fixing the weather data

# Extract only the weather data from the list after patching and fixing

weather <- list()

for (i in 1 : length(All_patched_fixed)){
  
  weather[[i]] <- All_patched_fixed[[i]][["weather"]]}


# Change the name of the elements of the list

names(weather) <- All_WS_90$Name


# Save the files in a separate folder

save_temperature_scenarios(weather, path = "./data/patched_temps", prefix = "WS")


# Compress the data 

zip(zipfile = "./data/patched_temps", files = "./data/patched_temps", flags = " a -idq -ep1",
    zip = "C:/Program Files/WinRAR/WinRAR")


# Remove the csv files

unlink("./data/patched_temps", recursive = T)


# Remove the files from the Argentinean folder since they are also included in the zip file

unlink("./data/Argentina", recursive = T)
